+++ src/factories/commands/trading-wallet/GetAddressCommandBuilder.js
@@ -1,12 +1,8 @@
 const logger = require('../../../logger')
-const PrivateKeyServiceBuilder = require('../../../../src/factories/PrivateKeyServiceBuilder')
 const GetAddressCommand = require('../../../commands/trading-wallet/GetAddressCommand')
 const GetAddressCommandValidator = require('../../../validators/commands/trading-wallet/GetAddressCommandValidator')
-const PrivateKeyValidator = require('../../../validators/PrivateKeyValidator')
 const TradingWalletServiceBuilder = require('../../TradingWalletServiceBuilder')
 
-const privateKeyService = PrivateKeyServiceBuilder.build()
-const privateKeyValidator = new PrivateKeyValidator(logger)
 const tradingWalletService = TradingWalletServiceBuilder.build()
 
 /**
@@ -15,8 +11,7 @@ const tradingWalletService = TradingWalletServiceBuilder.build()
 class GetAddressCommandBuilder {
   static build() {
     const getAddressCommandValidator = new GetAddressCommandValidator(logger)
-    const getAddressCommand = new GetAddressCommand(logger, tradingWalletService,
-      getAddressCommandValidator, privateKeyService, privateKeyValidator)
+    const getAddressCommand = new GetAddressCommand(logger, tradingWalletService, getAddressCommandValidator)
     return getAddressCommand
   }
 }